refactor dir stemcells location to cmds/stemcells
