+++ bin/ci
@@ -31,5 +31,5 @@
   ginkgo -r -p -v --noisyPendings main common stemcells
 
   echo -e "\n Vetting packages for potential issues..."
-  go tool vet main common stemcells
+  go tool vet main cmds common integration
 )
