+++ cmds/stemcells/light_stemcell_cmd_interface.go
