+++ cmds/stemcells/light_stemcell_vdi_cmd.go
