+++ cmds/stemcells/light_stemcell_vgbdgt_cmd.go
