+++ cmds/stemcells/stemcells_suite_test.go
